# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Livestream Studio API - OBS-like web application
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scenes/{sceneId}/sources:
    get:
      operationId: listSources
      tags: [sources]
      summary: List all sources in a scene
      parameters:
        - name: sceneId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of sources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Source"
    post:
      operationId: createSource
      tags: [sources]
      summary: Add a source to a scene
      parameters:
        - name: sceneId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SourceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"
  /sources/{id}:
    patch:
      operationId: updateSource
      tags: [sources]
      summary: Update a source
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SourceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"
        "404":
          description: Not found
    delete:
      operationId: deleteSource
      tags: [sources]
      summary: Delete a source
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /sources/{id}/layer:
    patch:
      operationId: updateSourceLayer
      tags: [sources]
      summary: Update source layer (z-index)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LayerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"

  # ── Stream ────────────────────────────────────────────────
  /stream/config:
    get:
      operationId: getStreamConfig
      tags: [stream]
      summary: Get stream configuration (RTMP URL + stream key)
      responses:
        "200":
          description: Stream configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamConfig"
    put:
      operationId: saveStreamConfig
      tags: [stream]
      summary: Save stream configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StreamConfigInput"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamConfig"
  /stream/start:
    post:
      operationId: startStream
      tags: [stream]
      summary: Start the livestream
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StreamStartInput"
      responses:
        "200":
          description: Stream started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamStatus"
        "400":
          description: Bad request
...
  /uploads:
    get:
      operationId: listUploads
      tags: [upload]
      summary: List all uploaded media assets
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [image, video, audio, overlay]
      responses:
        "200":
          description: List of uploads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MediaAsset"
  /uploads/{id}:
    delete:
      operationId: deleteUpload
      tags: [upload]
      summary: Delete an uploaded asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Stats (wow endpoints) ──────────────────────────────────
...
components:
  schemas:
...
    Source:
      type: object
      required: [id, sceneId, name, type, sortOrder, x, y, width, height, visible, locked, createdAt]
      properties:
        id:
          type: integer
        sceneId:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [camera, display, image, video, audio, browser, text, qrcode, clock, countdown, slideshow, videoPlaylist, audioPlaylist, pdf, logo, watermark]
        settings:
          type: object
          description: JSON blob of type-specific settings
          additionalProperties: true
        x:
          type: number
        y:
          type: number
        width:
          type: number
        height:
          type: number
        rotation:
          type: number
          default: 0
        opacity:
          type: number
          default: 1
        visible:
          type: boolean
          default: true
        locked:
          type: boolean
          default: false
        sortOrder:
          type: integer
        createdAt:
          type: string
    SourceInput:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
          enum: [camera, display, image, video, audio, browser, text, qrcode, clock, countdown, slideshow, videoPlaylist, audioPlaylist, pdf, logo, watermark]
        settings:
          type: object
          additionalProperties: true
        x:
          type: number
        y:
          type: number
        width:
          type: number
        height:
          type: number
        rotation:
          type: number
        opacity:
          type: number
    SourceUpdate:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [camera, display, image, video, audio, browser, text, qrcode, clock, countdown, slideshow, videoPlaylist, audioPlaylist, pdf, logo, watermark]
        settings:
          type: object
          additionalProperties: true
        x:
          type: number
        y:
          type: number
        width:
          type: number
        height:
          type: number
        rotation:
          type: number
        opacity:
          type: number
        visible:
          type: boolean
        locked:
          type: boolean
        sortOrder:
          type: integer
    LayerUpdate:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [bringToFront, bringForward, sendBackward, sendToBack]

    # ── Stream ────────────────────────────────────────────────
    StreamConfig:
      type: object
      required: [id, rtmpUrl, streamKey]
      properties:
        id:
          type: integer
        rtmpUrl:
          type: string
        streamKey:
          type: string
        platform:
          type: string
          enum: [facebook, youtube, twitch, custom]
    StreamConfigInput:
      type: object
      required: [rtmpUrl, streamKey]
      properties:
        rtmpUrl:
          type: string
          minLength: 1
        streamKey:
          type: string
          minLength: 1
        platform:
          type: string
          enum: [facebook, youtube, twitch, custom]
    StreamStartInput:
      type: object
      required: [rtmpUrl, streamKey]
      properties:
        rtmpUrl:
          type: string
          minLength: 1
        streamKey:
          type: string
          minLength: 1
    StreamStatus:
      type: object
      required: [state]
      properties:
        state:
          type: string
          enum: [idle, connecting, live, stopping, error, reconnecting]
        fps:
          type: ["number", "null"]
        bitrate:
          type: ["number", "null"]
        droppedFrames:
          type: ["integer", "null"]
        totalFrames:
          type: ["integer", "null"]
        uptimeSeconds:
          type: ["integer", "null"]
        cpuUsage:
          type: ["number", "null"]
        memoryMb:
          type: ["number", "null"]
        networkKbps:
          type: ["number", "null"]
        errorMessage:
          type: ["string", "null"]
...
    MediaAsset:
      type: object
      required: [id, filename, url, type, sizeBytes, createdAt]
      properties:
        id:
          type: integer
        filename:
          type: string
        url:
          type: string
        type:
          type: string
          enum: [image, video, audio, overlay]
        mimeType:
          type: string
        sizeBytes:
          type: integer
        createdAt:
          type: string