# lib/api-zod/src/generated/types/streamStatus.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Livestream Studio API - OBS-like web application
 * OpenAPI spec version: 0.1.0
 */
import type { StreamStatusState } from './streamStatusState';

export interface StreamStatus {
  state: StreamStatusState;
  /** @nullable */
  fps?: number | null;
  /** @nullable */
  bitrate?: number | null;
  /** @nullable */
  droppedFrames?: number | null;
  /** @nullable */
  totalFrames?: number | null;
  /** @nullable */
  uptimeSeconds?: number | null;
  /** @nullable */
  cpuUsage?: number | null;
  /** @nullable */
  memoryMb?: number | null;
  /** @nullable */
  networkKbps?: number | null;
  /** @nullable */
  errorMessage?: string | null;
}